Fix framework and library search paths
The linker/loader was configured to look in hardcoded directories that
seem to exist only on the core developers' machines, which is not of
much help to downstream users and generated a lot of unnecessary
warnings. This commit fixes those paths so they refer only to paths
relative to the project.